New structure for intra-`Ractor` communication.
This is similar to the proposed "basket" APIs in some ways: https://bugs.ruby-lang.org/issues/17298

Encapsulates:
- The query object (`String`, `Regexp`, `CYI`, `Pathname`, etc).
- The response (determined by CYO logic).
- A handle to the `Ractor` who should receive the response.

This solves my race-condition issues where several simultaneous requests could get mixed-up replies from the CYO area.
Since we're sending the `EverlastingMessage` with `move: true` we can just treat it as mutable, set the response,
and then forward the entire thing back to its intended destination. Right now the sender is always also the receiver
to implement the synchronous API, but this could just as easily be used to build a pipeline.